Api: Add Response For Delete Account

+++ app/Http/Controllers/api/v1/UserController.php
@@ -8,6 +8,7 @@ use Illuminate\Database\Eloquent\ { ModelNotFoundException };
 use Illuminate\Support\Facades\ { Storage, DB, Auth };
 use App\Http\Resources\v1\ { UserProfile, UserFullProfile, ProfileReportResource, MyProfile };
 use App\Http\Requests\Api\User\ { ProfileRequest, ProfileFilterRequest, ProfileReportRequest };
+use App\Http\Requests\Api\Authentication\ { DeleteAccountRequest };
 use App\Http\Requests\Api\General\ { PaginationRequest };
 use App\Models\ { User, Location, ProfileReport };
 
@@ -325,29 +326,33 @@ class UserController extends Controller
     // Delete Account
     public function deletAccount(Request $request)
     {
-        try{
-            $user = $request->user();
-            $user->delete();
-
-            return ([
-                'data'  =>  NULL,
-                'meta' => [
-                    'url'       =>  url()->current(),
-                    'api'       =>  $this->getVersion(),
-                    'language'  =>  app()->getLocale(),
-                    'message'   =>  trans('api.delete', ['entity' => __('Your Account')]),
-                ] ]);
-        } catch(ModelNotFoundException $exception) {                
-            switch ($exception->getModel()) {
-                case 'App\Models\User':
-                    $this->response['meta']['message'] = trans('api.not_found', ['entity' => __("User")]);
-                    break;
-                default:
-                    $this->response['meta']['message'] = trans('api.went_wrong');
-                    break;
-            };
-        } catch (\Exception $e) {
-            $this->storeErrorLog($e,'delete_account');
+        $rules = DeleteAccountRequest::rules();
+        if( $this->apiValidator($request->all(), $rules) ) {
+            try{
+                $user = $request->user();
+                $user->reason_of_delete = $request->reason;
+                $user->save(); 
+                $user->delete();
+                return ([
+                    'data'  =>  NULL,
+                    'meta' => [
+                        'url'       =>  url()->current(),
+                        'api'       =>  $this->getVersion(),
+                        'language'  =>  app()->getLocale(),
+                        'message'   =>  trans('api.delete', ['entity' => __('Your Account')]),
+                    ] ]);
+            } catch(ModelNotFoundException $exception) {                
+                switch ($exception->getModel()) {
+                    case 'App\Models\User':
+                        $this->response['meta']['message'] = trans('api.not_found', ['entity' => __("User")]);
+                        break;
+                    default:
+                        $this->response['meta']['message'] = trans('api.went_wrong');
+                        break;
+                };
+            } catch (\Exception $e) {
+                $this->storeErrorLog($e,'delete_account');
+            }
         }
         return $this->returnResponse();
     }
